feat: Удаление модели Order и связанных классов, добавление интеграционных тестов для процесса заказа и доставки
- Удалена модель Order и связанные классы, включая OrderQuerySet и Site, что упрощает структуру приложения.
- Добавлены интеграционные тесты для полного цикла заказ-доставка, включая создание и оплату заказов, создание посылок и доставок.
- Улучшены проверки и валидация в тестах, что повышает надежность и поддерживаемость системы.

Эти изменения способствуют упрощению кода и улучшению тестирования бизнес-логики.
