+++ atom/order/models/__init__.py
@@ -1,0 +1,6 @@
+"""Модели приложения orders."""
+
+from .order import Order
+from .site import Site
+
+__all__ = ["Order", "Site"]
