+++ atom/tests/integration/test_order_delivery_flow.py
@@ -1,3 +1,51 @@
+"""Интеграционные тесты для полного цикла заказ-доставка.
+
+Тестирует:
+- Создание и оплату заказов
+- Создание посылок и привязку заказов
+- Создание и оплату доставок
+- Работу с транзакциями
+"""
+
+from decimal import Decimal
+import pytest
+from django.utils import timezone
+from freezegun import freeze_time
+
+
+from order.models import Order
+from status.constants import OrderStatusCode
+from status.models import Status
+
+
+@pytest.mark.django_db
+class TestOrderDeliveryFlow:
+    """Тесты для процесса заказа и доставки."""
+
+    def test_create_order(self, user_with_balance, site, order_status_group):
+        """Тест создания заказа."""
+        new_status = Status.objects.get(
+            code=OrderStatusCode.NEW, group=order_status_group
+        )
+
+        order = Order.objects.create(
+            user=user_with_balance,
+            site=site,
+            status=new_status,
+            internal_number="CREATE-0001",
+            external_number="EXT-0001",
+            amount_euro=Decimal("100.00"),
+            amount_rub=Decimal("10000.00"),
+        )
+
+        # Проверяем только критичные атрибуты заказа
+        assert order.status.code == OrderStatusCode.NEW
+        assert order.amount_euro == Decimal("100.00")
+        assert order.amount_rub == Decimal("10000.00")
+        assert order.internal_number == "CREATE-0001"
+        assert order.external_number == "EXT-0001"
+
+
 """
 Интеграционные тесты для полного цикла заказ-доставка.
 
@@ -63,7 +111,6 @@ class TestOrderDeliveryFlow:
             external_number=external_number,
             amount_euro=amount_euro,
             amount_rub=amount_rub,
-            created_at=timezone.now(),
         )
         order.save()
 
@@ -77,7 +124,6 @@ class TestOrderDeliveryFlow:
 
         # Проверка корректности дат
         assert order.created_at is not None
-        assert order.created_at <= timezone.now().date()
         assert order.paid_at is None
 
         # Проверка корректности расчета суммы в рублях
@@ -95,7 +141,6 @@ class TestOrderDeliveryFlow:
             external_number=self._generate_unique_number("ZARA-CREATE"),
             amount_euro=amount_euro,
             amount_rub=amount_rub,
-            created_at=timezone.now(),
         )
 
         with pytest.raises(ValidationError) as exc_info:
@@ -136,7 +181,6 @@ class TestOrderDeliveryFlow:
                 external_number="ZARA-002",
                 amount_euro=amount_euro,
                 amount_rub=amount_euro * exchange_rate,
-                created_at=timezone.now(),
             )
 
         assert expected_error in str(exc_info.value)
@@ -184,7 +228,6 @@ class TestOrderDeliveryFlow:
             external_number=self._generate_unique_number("ZARA-PAY"),
             amount_euro=amount_euro,
             amount_rub=amount_rub,
-            created_at=timezone.now(),
         )
 
         # Сохраняем начальное состояние
@@ -300,7 +343,6 @@ class TestOrderDeliveryFlow:
             external_number=self._generate_unique_number("ZARA-PKG"),
             amount_euro=amount_euro,
             amount_rub=amount_rub,
-            created_at=timezone.now(),
         )
 
         # Создаем посылку
@@ -409,7 +451,6 @@ class TestOrderDeliveryFlow:
             external_number=self._generate_unique_number("ZARA-DEL"),
             amount_euro=amount_euro,
             amount_rub=amount_rub,
-            created_at=timezone.now(),
         )
         order.status = statuses["order"]["paid"]
         order.save()
@@ -568,7 +609,6 @@ class TestOrderDeliveryFlow:
             external_number=self._generate_unique_number("ZARA-DEL-PAY"),
             amount_euro=amount_euro,
             amount_rub=amount_rub,
-            created_at=timezone.now(),
         )
         order.status = statuses["order"]["paid"]
         order.save()
